Moving In progress Egg Math code to laptop

+++ include/types.h
@@ -10,7 +10,7 @@ typedef signed long long s64;
 
 typedef unsigned char u8;
 typedef unsigned short u16;
-typedef unsigned int u32;
+typedef unsigned long u32;
 typedef unsigned long long u64;
 
 #ifdef __CWCC__
